[level 2] Title: n^2 배열 자르기, Time: 43.96 ms, Memory: 23.9 MB -BaekjoonHub

+++ 프로그래머스/2/87390. n＾2 배열 자르기/n＾2 배열 자르기.cpp
@@ -8,8 +8,8 @@ vector<int> solution(int n, long long left, long long right) {
     vector<long long> vec;
 
     for(long long i = left; i<= right; i++){        
-        int x = i / n;
-        int y = i % n;
+        long long x = i / n;
+        long long y = i % n;
         
         if(x > y) {
             answer.push_back(x + 1);
